To Do App - updated

+++ To Do App/app.js
@@ -2,40 +2,48 @@ var list = document.getElementById("list");
 
 function addTodo() {
     var todo_item = document.getElementById("todo-item");
-    var li = document.createElement('li');
-    var listText = document.createTextNode(todo_item.value)
-    li.setAttribute('class', 'list-group-item')
-    li.appendChild(listText)
 
-    // delete Button
+    if (todo_item.value == "") {
+        alert('Please Insert Element');
+    } 
+    else {
 
-    var delBtn = document.createElement("button");
-    delBtn.setAttribute('class', 'btn btn-primary float-right ml-1');
-    delBtn.setAttribute('onclick', 'deleteItem(this)');
-    var icon = document.createElement("i");
-    icon.setAttribute('class', 'fa fa-trash');
-    icon.setAttribute('aria-hidden', 'true')
+        var li = document.createElement('li');
+        var listText = document.createTextNode(todo_item.value)
+        li.setAttribute('class', 'list-group-item')
+        li.appendChild(listText)
 
-    delBtn.appendChild(icon);
+        // delete Button
 
-    // edit btn
+        var delBtn = document.createElement("button");
+        delBtn.setAttribute('class', 'btn btn-primary float-right ml-1');
+        delBtn.setAttribute('onclick', 'deleteItem(this)');
+        var icon = document.createElement("i");
+        icon.setAttribute('class', 'fa fa-trash');
+        icon.setAttribute('aria-hidden', 'true')
 
-    var editBtn = document.createElement("button");
-    editBtn.setAttribute('class', 'btn btn-primary float-right ml-1');
-    editBtn.setAttribute('onclick', 'editItem(this)');
-    var editicon = document.createElement("i");
-    editicon.setAttribute('class', 'fa fa-pencil-square-o');
-    editicon.setAttribute('aria-hidden', 'true');
+        delBtn.appendChild(icon);
 
-    editBtn.appendChild(editicon);
-    delBtn.appendChild(icon);
+        // edit btn
 
-    li.appendChild(delBtn);
-    li.appendChild(editBtn)
+        var editBtn = document.createElement("button");
+        editBtn.setAttribute('class', 'btn btn-primary float-right ml-1');
+        editBtn.setAttribute('onclick', 'editItem(this)');
+        var editicon = document.createElement("i");
+        editicon.setAttribute('class', 'fa fa-pencil-square-o');
+        editicon.setAttribute('aria-hidden', 'true');
 
-    list.appendChild(li)
+        editBtn.appendChild(editicon);
+        delBtn.appendChild(icon);
 
-    todo_item.value = ""
+        li.appendChild(delBtn);
+        li.appendChild(editBtn)
+
+        list.appendChild(li)
+
+        todo_item.value = ""
+
+    }
 }
 
 
@@ -47,9 +55,9 @@ function deleteAll() {
     list.innerHTML = ""
 }
 
-function editItem(e){
-	var val = e.parentNode.firstChild.nodeValue;
-	var editValue = prompt("Enter the value",val);
-	e.parentNode.firstChild.nodeValue = editValue;
+function editItem(e) {
+    var val = e.parentNode.firstChild.nodeValue;
+    var editValue = prompt("Enter the value", val);
+    e.parentNode.firstChild.nodeValue = editValue;
 
 }
